requirements.txt
PyYAML>=5.3.1,<7.0
numpy>=1.21,<1.25
pandas>=1.3,<2.0
scipy>=1.7,<1.12